backend/requirements.txt
Flask==2.2.5
boto3==1.28.0
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Werkzeug==2.2.3
gunicorn==20.1.0
